pyproject.toml
[project]
name = "crawl4r"
version = "0.1.0"
description = "RAG ingestion pipeline for web crawling using LlamaIndex, TEI, Qwen3, and Qdrant"
requires-python = ">=3.10"
dependencies = [
    "llama-index-core>=0.14.0",
    "llama-index-vector-stores-qdrant",
    "llama-index-readers-file",
    "qdrant-client>=1.16.0",
    "watchdog>=6.0.0",
    "huggingface-hub",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv",
    "transformers>=4.57.5",
    "pyyaml>=6.0.0",
    "httpx==0.28.1",
    "fastapi>=0.128.0",
    "llama-index-cli>=0.5.3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "ty",
]

[tool.ruff]
line-length = 88
target-version = "py310"
cache-dir = ".cache/.ruff_cache"
exclude = ["scripts/**", ".cache/**", "build/**", "dist/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"examples/**" = ["E501"]
"tests/**" = ["E501"]

[tool.ty]

[tool.ty.environment]
python-version = "3.10"

[tool.ty.src]
include = ["crawl4r/**/*.py", "tests/**/*.py"]
exclude = [".cache/**", "build/**", "dist/**"]

[tool.ty.rules]
# Strict type checking rules
unresolved-attribute = "error"
unresolved-import = "error"
invalid-assignment = "error"
deprecated = "error"

[tool.pytest.ini_options]
cache_dir = ".cache/.pytest_cache"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
data_file = ".cache/.coverage"
omit = ["tests/*", ".cache/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "ty>=0.0.12",
    "respx>=0.21.0",
    "psutil>=6.1.0",
    "numpy>=2.2.1",
]
